requirements.txt
pymongo
dnspython
numpy
pandas
plotly